+++ algo/src/main/java/org/neo4j/graphalgo/MultiTypesProc.java
@@ -6,9 +6,7 @@ import org.neo4j.graphalgo.core.heavyweight.HeavyGraphFactory;
 import org.neo4j.graphalgo.impl.multiTypes.MultiTypes;
 import org.neo4j.graphalgo.results.ComputeAllMetaPathsResult;
 import org.neo4j.graphdb.GraphDatabaseService;
-import org.neo4j.kernel.api.KernelTransaction;
 import org.neo4j.kernel.internal.GraphDatabaseAPI;
-import org.neo4j.logging.Log;
 import org.neo4j.procedure.Context;
 import org.neo4j.procedure.Description;
 import org.neo4j.procedure.Name;
@@ -16,25 +14,19 @@ import org.neo4j.procedure.Procedure;
 
 public class MultiTypesProc {
 
-
     @Context
     public GraphDatabaseAPI api;
 
     @Context
     public GraphDatabaseService db;
 
-    @Context
-    public Log log;
-
-    @Context
-    public KernelTransaction transaction;
-
     @Procedure("algo.multiTypes")
-    @Description("CALL algo.multiTypes(edgeType:string) Does epic stuff.")
+    @Description("CALL algo.multiTypes(edgeType:string, typeLabel:string)" +
+            "- Convert a graph in that labels are nodes to which entities have an edge to " +
+            "to a graph where each node has their label in the label attribute.")
 
     public void multiTypes(
             @Name(value = "edgeType", defaultValue = "/type/object/type") String edgeType,
-            @Name(value = "entityLabel", defaultValue = "Entity") String entityLabel,
             @Name(value = "typeLabel", defaultValue = "Type") String typeLabel) throws Exception {
 
         final ComputeAllMetaPathsResult.Builder builder = ComputeAllMetaPathsResult.builder();
