+++ tests/src/test/java/org/neo4j/graphalgo/algo/MultiTypeTest.java
@@ -1,0 +1,66 @@
+package org.neo4j.graphalgo.algo;
+
+import org.junit.AfterClass;
+import org.junit.BeforeClass;
+import org.junit.Test;
+import org.neo4j.graphalgo.GettingStartedProc;
+import org.neo4j.graphalgo.TestDatabaseCreator;
+import org.neo4j.graphdb.Transaction;
+import org.neo4j.kernel.api.exceptions.KernelException;
+import org.neo4j.kernel.impl.proc.Procedures;
+import org.neo4j.kernel.internal.GraphDatabaseAPI;
+
+import static org.junit.Assert.assertTrue;
+
+public class MultiTypeTest {
+
+    private static GraphDatabaseAPI api;
+
+    @BeforeClass
+    public static void setup() throws KernelException {
+        final String cypher =
+                "CREATE (s:Node {name:'s', blow: 2})\n" +
+                        "CREATE (a:Node {name:'a'})\n" +
+                        "CREATE (b:Type {name:'b'})\n" +
+                        "CREATE (c:Type {name:'c'})\n" +
+
+                        "CREATE" +
+                        " (a)-[:OF_TYPE {cost:5, blue: 1}]->(b),\n" +
+                        " (a)-[:OF_TYPE {cost:10, blue: 1}]->(c),\n" +
+                        " (b)-[:OF_TYPE {cost:5, blue: 1}]->(c)";
+
+        api = TestDatabaseCreator.createTestDatabase();
+
+        api.getDependencyResolver()
+                .resolveDependency(Procedures.class)
+                .registerProcedure(GettingStartedProc.class);
+
+        try (Transaction tx = api.beginTx()) {
+            api.execute(cypher);
+            tx.success();
+        }
+    }
+
+    @AfterClass
+    public static void shutdownGraph() throws Exception {
+        api.shutdown();
+    }
+
+    @Test
+    public void testGettingStarted() throws Exception {
+
+        final String cypher = "CALL algo.multiTypes(\"OF_TYPE\", \"Type\")";
+
+        int i = 0;
+        for (Object o : api.getAllLabels())
+            i++;
+
+        api.execute(cypher);
+
+        int j = 0;
+        for (Object o : api.getAllLabels())
+            j++;
+
+        assertTrue(j > i);
+    }
+}
